rename stackoutput resource to tccpoutputs
