+++ service/controller/v25/resource/tccpoutputs/resource.go
@@ -1,4 +1,4 @@
-package stackoutput
+package tccpoutputs
 
 import (
 	"github.com/giantswarm/microerror"
@@ -6,7 +6,7 @@ import (
 )
 
 const (
-	Name = "stackoutputv25"
+	Name = "tccpoutputsv25"
 )
 
 type Config struct {
@@ -16,10 +16,10 @@ type Config struct {
 }
 
 // Resource implements an operatorkit resource and provides a mechanism to fetch
-// information from a cloud formation stack outputs of the Tenant Cluster Main
-// stack.
+// information from Cloud Formation stack outputs of the Tenant Cluster Control
+// Plane stack.
 //
-// The TCM manages the VPC Peering Connection. The peering connection ID is
+// The TCCP manages the VPC Peering Connection. The peering connection ID is
 // added to the controller context and used in the CPF stack.
 //
 type Resource struct {
