+++ service/controller/v25/resource/tccpoutputs/spec.go
@@ -1,4 +1,4 @@
-package stackoutput
+package tccpoutputs
 
 import "github.com/aws/aws-sdk-go/service/ec2"
